+++ client/lib/managementClient.ts
@@ -8,6 +8,9 @@ import * as debugModule from "debug";
 import { RequestResponseLink, createRequestResponseLink, sendRequest } from "./rpc";
 import { defaultLock } from "./util/utils";
 import { AmqpMessage } from ".";
+import { ConnectionContext } from "./connectionContext";
+import { ClientEntity } from "./clientEntity";
+import { translate } from "./errors";
 
 const debug = debugModule("azure:event-hubs:management");
 
@@ -65,26 +68,48 @@ export interface EventHubPartitionRuntimeInformation {
   type: "com.microsoft:partition";
 }
 
+export interface ManagementClientOptions {
+  address?: string;
+  audience?: string;
+}
+
 /**
  * @class ManagementClient
  * Descibes the EventHubs Management Client that talks
  * to the $management endpoint over AMQP connection.
  */
-export class ManagementClient {
+export class ManagementClient extends ClientEntity {
 
   readonly managementLock: string = `${Constants.managementRequestKey}-${uuid()}`;
   /**
+   * @property {string} entityPath - The name/path of the entity (hub name) for which the management
+   * request needs to be made.
+   */
+  entityPath: string;
+  /**
+   * @property {string} replyTo The reply to Guid for the management client.
+   */
+  replyTo: string = uuid();
+  /**
    * $management sender, receiver on the same session.
+   * @private
    */
   private _mgmtReqResLink?: RequestResponseLink;
 
   /**
    * @constructor
    * Instantiates the management client.
-   * @param entityPath - The name/path of the entity (hub name) for which the management request needs to be made.
+   * @param {BaseConnectionContext} context The connection context.
+   * @param {string} [address] The address for the management endpoint. For IotHub it will be
+   * `/messages/events/$management`.
    */
-  constructor(public entityPath: string) {
-    this.entityPath = entityPath;
+  constructor(context: ConnectionContext, options?: ManagementClientOptions) {
+    super(context, {
+      address: options && options.address ? options.address : Constants.management,
+      audience: options && options.audience ? options.audience : `${context.config.endpoint}${context.config.entityPath!}/$management`
+    });
+    this._context = context;
+    this.entityPath = context.config.entityPath as string;
   }
 
   /**
@@ -93,8 +118,8 @@ export class ManagementClient {
    * @param {Connection} connection - The established amqp connection
    * @returns {Promise<EventHubRuntimeInformation>}
    */
-  async getHubRuntimeInformation(connection: any): Promise<EventHubRuntimeInformation> {
-    const info: any = await this._makeManagementRequest(connection, Constants.eventHub);
+  async getHubRuntimeInformation(): Promise<EventHubRuntimeInformation> {
+    const info: any = await this._makeManagementRequest(Constants.eventHub);
     const runtimeInfo: EventHubRuntimeInformation = {
       path: info.name,
       createdAt: new Date(info.created_at),
@@ -102,7 +127,7 @@ export class ManagementClient {
       partitionIds: info.partition_ids,
       type: info.type
     };
-    debug("[%s] The hub runtime info is: %O", connection.options.id, runtimeInfo);
+    debug("[%s] The hub runtime info is: %O", this._context.connectionId, runtimeInfo);
     return runtimeInfo;
   }
 
@@ -112,8 +137,8 @@ export class ManagementClient {
    * @param {Connection} connection - The established amqp connection
    * @returns {Promise<Array<string>>}
    */
-  async getPartitionIds(connection: any): Promise<Array<string>> {
-    const runtimeInfo = await this.getHubRuntimeInformation(connection);
+  async getPartitionIds(): Promise<Array<string>> {
+    const runtimeInfo = await this.getHubRuntimeInformation();
     return runtimeInfo.partitionIds;
   }
 
@@ -123,11 +148,11 @@ export class ManagementClient {
    * @param {Connection} connection - The established amqp connection
    * @param {(string|number)} partitionId Partition ID for which partition information is required.
    */
-  async getPartitionInformation(connection: any, partitionId: string | number): Promise<EventHubPartitionRuntimeInformation> {
+  async getPartitionInformation(partitionId: string | number): Promise<EventHubPartitionRuntimeInformation> {
     if (!partitionId || (partitionId && typeof partitionId !== "string" && typeof partitionId !== "number")) {
       throw new Error("'partitionId' is a required parameter and must be of type: 'string' | 'number'.");
     }
-    const info: any = await this._makeManagementRequest(connection, Constants.partition, partitionId);
+    const info: any = await this._makeManagementRequest(Constants.partition, partitionId);
     const partitionInfo: EventHubPartitionRuntimeInformation = {
       beginningSequenceNumber: info.begin_sequence_number,
       hubPath: info.name,
@@ -137,7 +162,7 @@ export class ManagementClient {
       partitionId: info.partition,
       type: info.type
     };
-    debug("[%s] The partition info is: %O.", connection.options.id, partitionInfo);
+    debug("[%s] The partition info is: %O.", this._context.connectionId, partitionInfo);
     return partitionInfo;
   }
 
@@ -151,7 +176,9 @@ export class ManagementClient {
       if (this._mgmtReqResLink) {
         await rheaPromise.closeSession(this._mgmtReqResLink.session);
         debug("Successfully closed the management session.");
+        this._session = undefined;
         this._mgmtReqResLink = undefined;
+        clearTimeout(this._tokenRenewalTimer as NodeJS.Timer);
       }
     } catch (err) {
       const msg = `An error occurred while closing the management session: ${err}`;
@@ -160,13 +187,21 @@ export class ManagementClient {
     }
   }
 
-  private async _init(connection: any, endpoint: string, replyTo: string): Promise<void> {
+  private async _init(): Promise<void> {
     if (!this._mgmtReqResLink) {
-      const rxopt: rheaPromise.ReceiverOptions = { source: { address: endpoint }, name: replyTo, target: { address: replyTo } };
+      await this._negotiateClaim();
+      const rxopt: rheaPromise.ReceiverOptions = {
+        source: { address: this.address },
+        name: this.replyTo,
+        target: { address: this.replyTo }
+      };
+      const sropt: rheaPromise.SenderOptions = { target: { address: this.address } };
       debug("Creating a session for $management endpoint");
-      this._mgmtReqResLink = await createRequestResponseLink(connection, { target: { address: endpoint } }, rxopt);
+      this._mgmtReqResLink = await createRequestResponseLink(this._context.connection, sropt, rxopt);
+      this._session = this._mgmtReqResLink.session;
       debug("[%s] Created sender '%s' and receiver '%s' links for $management endpoint.",
-        connection.options.id, this._mgmtReqResLink.sender.name, this._mgmtReqResLink.receiver.name);
+        this._context.connectionId, this._mgmtReqResLink.sender.name, this._mgmtReqResLink.receiver.name);
+      await this._ensureTokenRenewal();
     }
   }
 
@@ -177,17 +212,15 @@ export class ManagementClient {
    * @param {string} type - The type of entity requested for. Valid values are "eventhub", "partition"
    * @param {string | number} [partitionId] - The partitionId. Required only when type is "partition".
    */
-  private async _makeManagementRequest(connection: any, type: "eventhub" | "partition", partitionId?: string | number): Promise<any> {
+  private async _makeManagementRequest(type: "eventhub" | "partition", partitionId?: string | number): Promise<any> {
     if (partitionId && typeof partitionId !== "string" && typeof partitionId !== "number") {
       throw new Error("'partitionId' is a required parameter and must be of type: 'string' | 'number'.");
     }
     try {
-      const endpoint = Constants.management;
-      const replyTo = uuid();
       const request: AmqpMessage = {
         body: Buffer.from(JSON.stringify([])),
         message_id: uuid(),
-        reply_to: replyTo,
+        reply_to: this.replyTo,
         application_properties: {
           operation: Constants.readOperation,
           name: this.entityPath as string,
@@ -197,9 +230,10 @@ export class ManagementClient {
       if (partitionId && type === Constants.partition) {
         request.application_properties!.partition = partitionId;
       }
-      await defaultLock.acquire(this.managementLock, () => { return this._init(connection, endpoint, replyTo); });
-      return sendRequest(connection, this._mgmtReqResLink!, request);
+      await defaultLock.acquire(this.managementLock, () => { return this._init(); });
+      return sendRequest(this._context.connection, this._mgmtReqResLink!, request);
     } catch (err) {
+      err = translate(err);
       debug("An error occurred while making the request to $management endpoint: %O", err);
       throw err;
     }